Refactoring composables into separate files for improved readability
